Add Queues and NLP
This commit adds the ability to publish events to queues along
with the handling of those jobs/messages.

This also adds NLP/Text Analysis to the system so that we can
see what our data is telling us and hopefully, in the future,
we can keep things that are trying to manipulate us away from us.